Remove geolocation from javascript and implement in python code.
I removed geolocation fetching from the javascript code and instead
chose to use a geolocation API, still from google, in python code.

This is actually better because it cuts down the number of requests
I need to make (only one set, instead of a big set for every client).

What I was actually doing wrong for the geolocation library was that I
wasn't requesting it with &libraries=geolocation or something when I was
loading the library. Geolocation with javascript is a library for the google
maps library so I needed to load it.

https://developers.google.com/maps/documentation/javascript/tutorial